refactor core planet path and drop ripple effects
