+++ client/pages/about.html
@@ -228,7 +228,6 @@
       </article>
     </main>
     <script defer src="../scripts/background.js"></script>
-    <script defer src="../scripts/ripple.js"></script>
     <script defer src="../scripts/nav.js"></script>
   </body>
 </html>
