+++ client/pages/abyss.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-abyss" style="--realm-color:#bf9dff;--ripple-color:#bf9dff"><h1>Abyss</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Silent Chase</button><div class="cluster-menu"><h2>Nocturnis</h2><div class="core-emotion">fear</div><hr class="cluster-divider"/><ul><li>anxiety</li><li>suspicion</li><li>dread</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Hollow Veil</button><div class="cluster-menu"><h2>Umbrak</h2><div class="core-emotion">emptiness</div><hr class="cluster-divider"/><ul><li>numbness</li><li>dysregulation</li><li>isolated</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Frozen Nerve</button><div class="cluster-menu"><h2>Nerveris</h2><div class="core-emotion">panic</div><hr class="cluster-divider"/><ul><li>disoriented</li><li>overwhelmed</li><li>paranoia</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Bound Breath</button><div class="cluster-menu"><h2>Thalix</h2><div class="core-emotion">helplessness</div><hr class="cluster-divider"/><ul><li>collapse</li><li>suffocation</li><li>powerlessness</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Null Horizon</button><div class="cluster-menu"><h2>Abyzeth</h2><div class="core-emotion">hopelessness</div><hr class="cluster-divider"/><ul><li>despair</li><li>derealization</li><li>meaninglessness</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🕳️</div><h2>Realm Space</h2><button class="select-btn">Visit Abyss Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-abyss" style="--realm-color:#bf9dff"><h1>Abyss</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Silent Chase</button><div class="cluster-menu"><h2>Nocturnis</h2><div class="core-emotion">fear</div><hr class="cluster-divider"/><ul><li>anxiety</li><li>suspicion</li><li>dread</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Hollow Veil</button><div class="cluster-menu"><h2>Umbrak</h2><div class="core-emotion">emptiness</div><hr class="cluster-divider"/><ul><li>numbness</li><li>dysregulation</li><li>isolated</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Frozen Nerve</button><div class="cluster-menu"><h2>Nerveris</h2><div class="core-emotion">panic</div><hr class="cluster-divider"/><ul><li>disoriented</li><li>overwhelmed</li><li>paranoia</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Bound Breath</button><div class="cluster-menu"><h2>Thalix</h2><div class="core-emotion">helplessness</div><hr class="cluster-divider"/><ul><li>collapse</li><li>suffocation</li><li>powerlessness</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Null Horizon</button><div class="cluster-menu"><h2>Abyzeth</h2><div class="core-emotion">hopelessness</div><hr class="cluster-divider"/><ul><li>despair</li><li>derealization</li><li>meaninglessness</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🕳️</div><h2>Realm Space</h2><button class="select-btn">Visit Abyss Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
