+++ client/pages/cavern.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-cavern" style="--realm-color:#9ef2e3;--ripple-color:#9ef2e3"><h1>Cavern</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">False Shrine</button><div class="cluster-menu"><h2>Bedesto</h2><div class="core-emotion">idolization</div><hr class="cluster-divider"/><ul><li>obsession</li><li>projection</li><li>dependence</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Smoky Mirror</button><div class="cluster-menu"><h2>Thirsa</h2><div class="core-emotion">envy</div><hr class="cluster-divider"/><ul><li>inferiority</li><li>superiority</li><li>imposter syndrome</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Venom Hold</button><div class="cluster-menu"><h2>Kerelos</h2><div class="core-emotion">jealousy</div><hr class="cluster-divider"/><ul><li>insecure</li><li>possessive</li><li>clingy</li><li>territorial</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Itchy Bite</button><div class="cluster-menu"><h2>Varkir</h2><div class="core-emotion">bitterness</div><hr class="cluster-divider"/><ul><li>resentment</li><li>spite</li><li>pettiness</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Spiky Throne</button><div class="cluster-menu"><h2>Tharnis</h2><div class="core-emotion">contempt</div><hr class="cluster-divider"/><ul><li>entitlement</li><li>judgmental</li><li>smug</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🪨</div><h2>Realm Space</h2><button class="select-btn">Visit Cavern Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-cavern" style="--realm-color:#9ef2e3"><h1>Cavern</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">False Shrine</button><div class="cluster-menu"><h2>Bedesto</h2><div class="core-emotion">idolization</div><hr class="cluster-divider"/><ul><li>obsession</li><li>projection</li><li>dependence</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Smoky Mirror</button><div class="cluster-menu"><h2>Thirsa</h2><div class="core-emotion">envy</div><hr class="cluster-divider"/><ul><li>inferiority</li><li>superiority</li><li>imposter syndrome</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Venom Hold</button><div class="cluster-menu"><h2>Kerelos</h2><div class="core-emotion">jealousy</div><hr class="cluster-divider"/><ul><li>insecure</li><li>possessive</li><li>clingy</li><li>territorial</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Itchy Bite</button><div class="cluster-menu"><h2>Varkir</h2><div class="core-emotion">bitterness</div><hr class="cluster-divider"/><ul><li>resentment</li><li>spite</li><li>pettiness</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Spiky Throne</button><div class="cluster-menu"><h2>Tharnis</h2><div class="core-emotion">contempt</div><hr class="cluster-divider"/><ul><li>entitlement</li><li>judgmental</li><li>smug</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🪨</div><h2>Realm Space</h2><button class="select-btn">Visit Cavern Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
