+++ client/pages/dross.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-dross" style="--realm-color:#f0c9b3;--ripple-color:#f0c9b3"><h1>Dross</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Putrid Force</button><div class="cluster-menu"><h2>Graskul</h2><div class="core-emotion">disgust</div><hr class="cluster-divider"/><ul><li>repulsion</li><li>revulsion</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Tarry Bone</button><div class="cluster-menu"><h2>Exulith</h2><div class="core-emotion">self-loathing</div><hr class="cluster-divider"/><ul><li>unclean</li><li>stained</li><li>contaminated</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Stolen Doll</button><div class="cluster-menu"><h2>Trocita</h2><div class="core-emotion">violated</div><hr class="cluster-divider"/><ul><li>dehumanized</li><li>fetishized</li><li>used</li><li>exploited</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Pale Shiver</button><div class="cluster-menu"><h2>Ofoki</h2><div class="core-emotion">malaise</div><hr class="cluster-divider"/><ul><li>nauseated</li><li>itchy</li><li>sweaty</li><li>stinky</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">☣️</div><h2>Realm Space</h2><button class="select-btn">Visit Dross Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-dross" style="--realm-color:#f0c9b3"><h1>Dross</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Putrid Force</button><div class="cluster-menu"><h2>Graskul</h2><div class="core-emotion">disgust</div><hr class="cluster-divider"/><ul><li>repulsion</li><li>revulsion</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Tarry Bone</button><div class="cluster-menu"><h2>Exulith</h2><div class="core-emotion">self-loathing</div><hr class="cluster-divider"/><ul><li>unclean</li><li>stained</li><li>contaminated</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Stolen Doll</button><div class="cluster-menu"><h2>Trocita</h2><div class="core-emotion">violated</div><hr class="cluster-divider"/><ul><li>dehumanized</li><li>fetishized</li><li>used</li><li>exploited</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Pale Shiver</button><div class="cluster-menu"><h2>Ofoki</h2><div class="core-emotion">malaise</div><hr class="cluster-divider"/><ul><li>nauseated</li><li>itchy</li><li>sweaty</li><li>stinky</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">☣️</div><h2>Realm Space</h2><button class="select-btn">Visit Dross Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
