+++ client/pages/ember.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-ember" style="--realm-color:#ffb194;--ripple-color:#ffb194"><h1>Ember</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Infernal Mammoth</button><div class="cluster-menu"><h2>Fureza</h2><div class="core-emotion">rage</div><hr class="cluster-divider"/><ul><li>indignant</li><li>berserk</li><li>smoldering</li><li>destructive</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Brazen Rhino</button><div class="cluster-menu"><h2>Romana</h2><div class="core-emotion">reckless</div><hr class="cluster-divider"/><ul><li>audacity</li><li>thrill-seeking</li><li>impulsive</li><li>daring</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Noisy Wasp</button><div class="cluster-menu"><h2>Agnera</h2><div class="core-emotion">frustrated</div><hr class="cluster-divider"/><ul><li>irritated</li><li>impatient</li><li>agitated</li><li>overthinking</li><li>hypervigilant</li><li>restless</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Vicious Cobra</button><div class="cluster-menu"><h2>Provota</h2><div class="core-emotion">aggressive</div><hr class="cluster-divider"/><ul><li>confrontational</li><li>defiant</li><li>hostile</li><li>provocative</li><li>rebellious</li><li>combative</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🔥</div><h2>Realm Space</h2><button class="select-btn">Visit Ember Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-ember" style="--realm-color:#ffb194"><h1>Ember</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Infernal Mammoth</button><div class="cluster-menu"><h2>Fureza</h2><div class="core-emotion">rage</div><hr class="cluster-divider"/><ul><li>indignant</li><li>berserk</li><li>smoldering</li><li>destructive</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Brazen Rhino</button><div class="cluster-menu"><h2>Romana</h2><div class="core-emotion">reckless</div><hr class="cluster-divider"/><ul><li>audacity</li><li>thrill-seeking</li><li>impulsive</li><li>daring</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Noisy Wasp</button><div class="cluster-menu"><h2>Agnera</h2><div class="core-emotion">frustrated</div><hr class="cluster-divider"/><ul><li>irritated</li><li>impatient</li><li>agitated</li><li>overthinking</li><li>hypervigilant</li><li>restless</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Vicious Cobra</button><div class="cluster-menu"><h2>Provota</h2><div class="core-emotion">aggressive</div><hr class="cluster-divider"/><ul><li>confrontational</li><li>defiant</li><li>hostile</li><li>provocative</li><li>rebellious</li><li>combative</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🔥</div><h2>Realm Space</h2><button class="select-btn">Visit Ember Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
