+++ client/pages/explore.html
@@ -88,7 +88,6 @@
       </section>
     </main>
     <script defer src="../scripts/background.js"></script>
-    <script defer src="../scripts/ripple.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/explore.js"></script>
   </body>
