+++ client/pages/glare.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-glare" style="--realm-color:#ffd96a;--ripple-color:#ffd96a"><h1>Glare</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Fallen Eye</button><div class="cluster-menu"><h2>Rassembar</h2><div class="core-emotion">shame</div><hr class="cluster-divider"/><ul><li>humiliation</li><li>exposure</li><li>unworthiness</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Deadlight Hall</button><div class="cluster-menu"><h2>Censyr</h2><div class="core-emotion">repressed</div><hr class="cluster-divider"/><ul><li>tension</li><li>avoidance</li><li>self-denial</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Raucous Laugh</button><div class="cluster-menu"><h2>Tawnter</h2><div class="core-emotion">judged</div><hr class="cluster-divider"/><ul><li>mocked</li><li>spotlighted</li><li>accused</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">👁️</div><h2>Realm Space</h2><button class="select-btn">Visit Glare Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-glare" style="--realm-color:#ffd96a"><h1>Glare</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Fallen Eye</button><div class="cluster-menu"><h2>Rassembar</h2><div class="core-emotion">shame</div><hr class="cluster-divider"/><ul><li>humiliation</li><li>exposure</li><li>unworthiness</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Deadlight Hall</button><div class="cluster-menu"><h2>Censyr</h2><div class="core-emotion">repressed</div><hr class="cluster-divider"/><ul><li>tension</li><li>avoidance</li><li>self-denial</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Raucous Laugh</button><div class="cluster-menu"><h2>Tawnter</h2><div class="core-emotion">judged</div><hr class="cluster-divider"/><ul><li>mocked</li><li>spotlighted</li><li>accused</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">👁️</div><h2>Realm Space</h2><button class="select-btn">Visit Glare Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
