+++ client/pages/home.html
@@ -115,7 +115,6 @@
       </section>
     </main>
     <script defer src="../scripts/background.js"></script>
-    <script defer src="../scripts/ripple.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/checkin.js"></script>
   </body>
