+++ client/pages/languish.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-languish" style="--realm-color:#86bfff;--ripple-color:#86bfff"><h1>Languish</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Glass Crypt</button><div class="cluster-menu"><h2>Dolenza</h2><div class="core-emotion">despair</div><hr class="cluster-divider"/><ul><li>hopelessness</li><li>exhaustion</li><li>burnout</li><li>resignation</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Blank Prism</button><div class="cluster-menu"><h2>Sedra</h2><div class="core-emotion">numbness</div><hr class="cluster-divider"/><ul><li>lost</li><li>disconnected</li><li>detached</li><li>shut down</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Sapphire Chamber</button><div class="cluster-menu"><h2>Trosta</h2><div class="core-emotion">grief</div><hr class="cluster-divider"/><ul><li>sorrow</li><li>mourning</li><li>heartache</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Empty House</button><div class="cluster-menu"><h2>Estrana</h2><div class="core-emotion">isolation</div><hr class="cluster-divider"/><ul><li>unreachable</li><li>abandoned</li><li>unseen</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">💧</div><h2>Realm Space</h2><button class="select-btn">Visit Languish Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-languish" style="--realm-color:#86bfff"><h1>Languish</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Glass Crypt</button><div class="cluster-menu"><h2>Dolenza</h2><div class="core-emotion">despair</div><hr class="cluster-divider"/><ul><li>hopelessness</li><li>exhaustion</li><li>burnout</li><li>resignation</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Blank Prism</button><div class="cluster-menu"><h2>Sedra</h2><div class="core-emotion">numbness</div><hr class="cluster-divider"/><ul><li>lost</li><li>disconnected</li><li>detached</li><li>shut down</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Sapphire Chamber</button><div class="cluster-menu"><h2>Trosta</h2><div class="core-emotion">grief</div><hr class="cluster-divider"/><ul><li>sorrow</li><li>mourning</li><li>heartache</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Empty House</button><div class="cluster-menu"><h2>Estrana</h2><div class="core-emotion">isolation</div><hr class="cluster-divider"/><ul><li>unreachable</li><li>abandoned</li><li>unseen</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">💧</div><h2>Realm Space</h2><button class="select-btn">Visit Languish Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
