+++ client/pages/mist.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-mist" style="--realm-color:#cfe8f8;--ripple-color:#cfe8f8"><h1>Mist</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Lucid Thread</button><div class="cluster-menu"><h2>Verio</h2><div class="core-emotion">curious</div><hr class="cluster-divider"/><ul><li>intrigue</li><li>spark</li><li>suspense</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Ancient Quiet</button><div class="cluster-menu"><h2>Zonder</h2><div class="core-emotion">awe</div><hr class="cluster-divider"/><ul><li>wonder</li><li>reverence</li><li>timelessness</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Fractal Maze</button><div class="cluster-menu"><h2>Obsceris</h2><div class="core-emotion">confusion</div><hr class="cluster-divider"/><ul><li>ambiguity</li><li>unease</li><li>lost</li><li>disoriented</li><li>bewildered</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Shallow Anchor</button><div class="cluster-menu"><h2>Dysira</h2><div class="core-emotion">doubt</div><hr class="cluster-divider"/><ul><li>skeptical</li><li>suspicious</li><li>distrust</li><li>second guessing</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Myriad Glitch</button><div class="cluster-menu"><h2>Umbranova</h2><div class="core-emotion">surreal</div><hr class="cluster-divider"/><ul><li>uncanny</li><li>absurd</li><li>nonsensical</li><li>paradoxical</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🌫️</div><h2>Realm Space</h2><button class="select-btn">Visit Mist Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-mist" style="--realm-color:#cfe8f8"><h1>Mist</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Lucid Thread</button><div class="cluster-menu"><h2>Verio</h2><div class="core-emotion">curious</div><hr class="cluster-divider"/><ul><li>intrigue</li><li>spark</li><li>suspense</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Ancient Quiet</button><div class="cluster-menu"><h2>Zonder</h2><div class="core-emotion">awe</div><hr class="cluster-divider"/><ul><li>wonder</li><li>reverence</li><li>timelessness</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Fractal Maze</button><div class="cluster-menu"><h2>Obsceris</h2><div class="core-emotion">confusion</div><hr class="cluster-divider"/><ul><li>ambiguity</li><li>unease</li><li>lost</li><li>disoriented</li><li>bewildered</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Shallow Anchor</button><div class="cluster-menu"><h2>Dysira</h2><div class="core-emotion">doubt</div><hr class="cluster-divider"/><ul><li>skeptical</li><li>suspicious</li><li>distrust</li><li>second guessing</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Myriad Glitch</button><div class="cluster-menu"><h2>Umbranova</h2><div class="core-emotion">surreal</div><hr class="cluster-divider"/><ul><li>uncanny</li><li>absurd</li><li>nonsensical</li><li>paradoxical</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🌫️</div><h2>Realm Space</h2><button class="select-btn">Visit Mist Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
