+++ client/pages/oasis.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-oasis" style="--realm-color:#76ffe5;--ripple-color:#76ffe5"><h1>Oasis</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Verdant Hearth</button><div class="cluster-menu"><h2>Merulo</h2><div class="core-emotion">safe</div><hr class="cluster-divider"/><ul><li>hydrated</li><li>rested</li><li>regulated</li><li>full</li><li>healthy</li><li>content</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Radiant Grove</button><div class="cluster-menu"><h2>Solene</h2><div class="core-emotion">joy</div><hr class="cluster-divider"/><ul><li>delighted</li><li>uplifted</li><li>beaming</li><li>radiant</li><li>ecstatic</li><li>excited</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Honey Bloom</button><div class="cluster-menu"><h2>Lubovu</h2><div class="core-emotion">love</div><hr class="cluster-divider"/><ul><li>affectionate</li><li>compassionate</li><li>nurtured</li><li>endearing</li><li>tender</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Open Palm</button><div class="cluster-menu"><h2>Gawyn</h2><div class="core-emotion">grateful</div><hr class="cluster-divider"/><ul><li>fortunate</li><li>acknowledged</li><li>cherished</li><li>accepted</li><li>received</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Woven Circle</button><div class="cluster-menu"><h2>Allora</h2><div class="core-emotion">belonging</div><hr class="cluster-divider"/><ul><li>centered</li><li>rooted</li><li>anchored</li><li>grounded</li><li>connected</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Twinkle Toe</button><div class="cluster-menu"><h2>Vivaly</h2><div class="core-emotion">playful</div><hr class="cluster-divider"/><ul><li>carefree</li><li>lighthearted</li><li>amused</li><li>silly</li><li>giddy</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Soft Meadow</button><div class="cluster-menu"><h2>Navari</h2><div class="core-emotion">peace</div><hr class="cluster-divider"/><ul><li>serene</li><li>tranquil</li><li>ease</li><li>untroubled</li><li>calm</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🌴</div><h2>Realm Space</h2><button class="select-btn">Visit Oasis Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-oasis" style="--realm-color:#76ffe5"><h1>Oasis</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Verdant Hearth</button><div class="cluster-menu"><h2>Merulo</h2><div class="core-emotion">safe</div><hr class="cluster-divider"/><ul><li>hydrated</li><li>rested</li><li>regulated</li><li>full</li><li>healthy</li><li>content</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Radiant Grove</button><div class="cluster-menu"><h2>Solene</h2><div class="core-emotion">joy</div><hr class="cluster-divider"/><ul><li>delighted</li><li>uplifted</li><li>beaming</li><li>radiant</li><li>ecstatic</li><li>excited</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Honey Bloom</button><div class="cluster-menu"><h2>Lubovu</h2><div class="core-emotion">love</div><hr class="cluster-divider"/><ul><li>affectionate</li><li>compassionate</li><li>nurtured</li><li>endearing</li><li>tender</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Open Palm</button><div class="cluster-menu"><h2>Gawyn</h2><div class="core-emotion">grateful</div><hr class="cluster-divider"/><ul><li>fortunate</li><li>acknowledged</li><li>cherished</li><li>accepted</li><li>received</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Woven Circle</button><div class="cluster-menu"><h2>Allora</h2><div class="core-emotion">belonging</div><hr class="cluster-divider"/><ul><li>centered</li><li>rooted</li><li>anchored</li><li>grounded</li><li>connected</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Twinkle Toe</button><div class="cluster-menu"><h2>Vivaly</h2><div class="core-emotion">playful</div><hr class="cluster-divider"/><ul><li>carefree</li><li>lighthearted</li><li>amused</li><li>silly</li><li>giddy</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Soft Meadow</button><div class="cluster-menu"><h2>Navari</h2><div class="core-emotion">peace</div><hr class="cluster-divider"/><ul><li>serene</li><li>tranquil</li><li>ease</li><li>untroubled</li><li>calm</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🌴</div><h2>Realm Space</h2><button class="select-btn">Visit Oasis Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
