+++ client/pages/profile.html
@@ -64,7 +64,6 @@
       </section>
     </main>
     <script defer src="../scripts/background.js"></script>
-    <script defer src="../scripts/ripple.js"></script>
     <script defer src="../scripts/nav.js"></script>
   </body>
 </html>
