+++ client/pages/trace.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-trace" style="--realm-color:#ffa3c0;--ripple-color:#ffa3c0"><h1>Trace</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Sepia Garden</button><div class="cluster-menu"><h2>Yorell</h2><div class="core-emotion">nostalgia</div><hr class="cluster-divider"/><ul><li>sentimental</li><li>familiarity</li><li>reminiscing</li><li>reflecting</li><li>pondering</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Word Cemetery</button><div class="cluster-menu"><h2>Renmor</h2><div class="core-emotion">regret</div><hr class="cluster-divider"/><ul><li>guilt</li><li>incompletion</li><li>lingering</li><li>aching</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Aurora Shadow</button><div class="cluster-menu"><h2>Fenric</h2><div class="core-emotion">yearning</div><hr class="cluster-divider"/><ul><li>craving</li><li>wistful</li><li>bittersweet</li><li>homesick</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Butterfly Canopy</button><div class="cluster-menu"><h2>Drisano</h2><div class="core-emotion">daydreaming</div><hr class="cluster-divider"/><ul><li>drifting</li><li>imagining</li><li>musing</li><li>contemplating</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🌀</div><h2>Realm Space</h2><button class="select-btn">Visit Trace Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-trace" style="--realm-color:#ffa3c0"><h1>Trace</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Sepia Garden</button><div class="cluster-menu"><h2>Yorell</h2><div class="core-emotion">nostalgia</div><hr class="cluster-divider"/><ul><li>sentimental</li><li>familiarity</li><li>reminiscing</li><li>reflecting</li><li>pondering</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Word Cemetery</button><div class="cluster-menu"><h2>Renmor</h2><div class="core-emotion">regret</div><hr class="cluster-divider"/><ul><li>guilt</li><li>incompletion</li><li>lingering</li><li>aching</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Aurora Shadow</button><div class="cluster-menu"><h2>Fenric</h2><div class="core-emotion">yearning</div><hr class="cluster-divider"/><ul><li>craving</li><li>wistful</li><li>bittersweet</li><li>homesick</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Butterfly Canopy</button><div class="cluster-menu"><h2>Drisano</h2><div class="core-emotion">daydreaming</div><hr class="cluster-divider"/><ul><li>drifting</li><li>imagining</li><li>musing</li><li>contemplating</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🌀</div><h2>Realm Space</h2><button class="select-btn">Visit Trace Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
