+++ client/pages/universe.html
@@ -62,7 +62,6 @@
       </div>
     </div>
     <script defer src="../scripts/background.js"></script>
-    <script defer src="../scripts/ripple.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/overlayData.js" type="module"></script>
     <script src="../scripts/universe.js" defer></script>
