+++ client/pages/zenith.html
@@ -15,7 +15,7 @@
     </div>
     <div id="stars"></div>
     <nav id="main-nav" aria-label="primary"></nav>
-    <main id="react-root"><div class="realm realm-zenith" style="--realm-color:#ffeba3;--ripple-color:#ffeba3"><h1>Zenith</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Steadfast Anvil</button><div class="cluster-menu"><h2>Milenios</h2><div class="core-emotion">endurance</div><hr class="cluster-divider"/><ul><li>resilient</li><li>disciplined</li><li>fortified</li><li>resolve</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Primordial Pillar</button><div class="cluster-menu"><h2>Eladon</h2><div class="core-emotion">pride</div><hr class="cluster-divider"/><ul><li>dignity</li><li>integrity</li><li>trustworthy</li><li>reputable</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Luminescent Crown</button><div class="cluster-menu"><h2>Kalyra</h2><div class="core-emotion">empowered</div><hr class="cluster-divider"/><ul><li>confident</li><li>inspired</li><li>triumphant</li><li>strong-willed</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Omega Threshold</button><div class="cluster-menu"><h2>Zepharel</h2><div class="core-emotion">liberation</div><hr class="cluster-divider"/><ul><li>elevation</li><li>release</li><li>awakening</li><li>eureka</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🚀</div><h2>Realm Space</h2><button class="select-btn">Visit Zenith Space</button></div></section></div></main>
+    <main id="react-root"><div class="realm realm-zenith" style="--realm-color:#ffeba3"><h1>Zenith</h1><section id="realm-space" class="realm-section"><div class="cluster-bubbles"><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Steadfast Anvil</button><div class="cluster-menu"><h2>Milenios</h2><div class="core-emotion">endurance</div><hr class="cluster-divider"/><ul><li>resilient</li><li>disciplined</li><li>fortified</li><li>resolve</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Primordial Pillar</button><div class="cluster-menu"><h2>Eladon</h2><div class="core-emotion">pride</div><hr class="cluster-divider"/><ul><li>dignity</li><li>integrity</li><li>trustworthy</li><li>reputable</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Luminescent Crown</button><div class="cluster-menu"><h2>Kalyra</h2><div class="core-emotion">empowered</div><hr class="cluster-divider"/><ul><li>confident</li><li>inspired</li><li>triumphant</li><li>strong-willed</li></ul></div></div><div class="cluster"><button type="button" class="cluster-bubble" aria-expanded="false">Omega Threshold</button><div class="cluster-menu"><h2>Zepharel</h2><div class="core-emotion">liberation</div><hr class="cluster-divider"/><ul><li>elevation</li><li>release</li><li>awakening</li><li>eureka</li></ul></div></div></div><hr class="realm-divider"/><div class="realm-bottom"><div class="realm-icon">🚀</div><h2>Realm Space</h2><button class="select-btn">Visit Zenith Space</button></div></section></div></main>
     <script defer src="../scripts/background.js"></script>
     <script defer src="../scripts/nav.js"></script>
     <script defer src="../scripts/clusters.js"></script>
