+++ src/data/corePlanets/cavern.ts
