+++ src/data/corePlanets/languish.ts
