+++ src/data/index.ts
@@ -29,13 +29,13 @@ export { clusters as traceClusters } from './clusters/trace'
 export { clusters as zenithClusters } from './clusters/zenith'
 
 // Planet data
-export { corePlanets as abyssPlanets } from './planets/abyss'
-export { corePlanets as cavernPlanets } from './planets/cavern'
-export { corePlanets as drossPlanets } from './planets/dross'
-export { corePlanets as emberPlanets } from './planets/ember'
-export { corePlanets as glarePlanets } from './planets/glare'
-export { corePlanets as languishPlanets } from './planets/languish'
-export { corePlanets as mistPlanets } from './planets/mist'
-export { corePlanets as oasisPlanets } from './planets/oasis'
-export { corePlanets as tracePlanets } from './planets/trace'
-export { corePlanets as zenithPlanets } from './planets/zenith'
+export { corePlanets as abyssPlanets } from './corePlanets/abyss'
+export { corePlanets as cavernPlanets } from './corePlanets/cavern'
+export { corePlanets as drossPlanets } from './corePlanets/dross'
+export { corePlanets as emberPlanets } from './corePlanets/ember'
+export { corePlanets as glarePlanets } from './corePlanets/glare'
+export { corePlanets as languishPlanets } from './corePlanets/languish'
+export { corePlanets as mistPlanets } from './corePlanets/mist'
+export { corePlanets as oasisPlanets } from './corePlanets/oasis'
+export { corePlanets as tracePlanets } from './corePlanets/trace'
+export { corePlanets as zenithPlanets } from './corePlanets/zenith'
